// src/app/core/features/Faculty-of-Archaeology/model/unit.model.ts
// src/app/model/unit.model.ts

export interface UnitGoal {
  id: string;
  index: number;
  goalName: string;
  aboutId: string;
}

export interface UnitAttachment {
  id: string;
  fileName: string;
  isPublic: boolean;
  relativePath: string;
  folderName: string;
  url: string;
  unitId: string;
}

export interface Unit {
  id: string;
  pageId: string;
  unitTitle: string;
  unitTitleEn: string;
  aboutId: string;
  content: string;
  mission: string;
  vision: string;
  history: string | null;
  goals: UnitGoal[];
  unitAttachments: UnitAttachment[];
}

export interface UnitDetail {
  id: string;
  title: string;
  content: string;
  unitPlace: string;
  unitId: string;
  unitTitle: string;
  unitAttachments: UnitAttachment[];
}

export interface UnitMember {
  id: string;
  isLeader: boolean;
  unitId: string;
  unitTitle: string;
  memberId: string;
  memberName: string;
}

export interface UnitSection {
  id: string;
  title: string;
  route: string;
  icon?: string;
}